+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update capitalization cleanup postprocessing to support sentences
   ending before quotation marks. For instance, `"Test." test` will now
   correct to `"Test." Test` where previously it would not.
+* Fix visual line breaks when backslash is followed by whitespace
 
 ### 0.1.9
 * Update experimental in-eval ref lookup. Now exposed by two